requirements.txt
python-telegram-bot==20.3
beautifulsoup4==4.12.2
requests==2.28.2